+++ src/plugins/extra/backgrounds/Unsplash/Unsplash.tsx
@@ -10,7 +10,7 @@ interface Props extends Settings {
   darken: boolean;
   focus: boolean;
   local: State;
-  setLocal: (state: State) => void;
+  setLocal: (state: Partial<State>) => void;
 }
 
 interface State {
@@ -20,30 +20,29 @@ interface State {
 
 class Unsplash extends React.PureComponent<Props, State> {
   static defaultProps: Partial<Props> = defaultProps;
-
   state: State = {};
-
-  private debouncedRefresh = debounce(this.refresh, 500);
+  private refreshDebounced = debounce(this.refresh, 500);
 
   componentWillMount() {
     // Fetch or pull from cache current image
-    if (this.props.local && this.props.local.next && this.props.local.next.data) {
-      this.set(this.props.local.next);
+    if (this.props.local.next && this.props.local.next.data) {
+      this.setImage(this.props.local.next);
     } else {
-      this.fetch(this.props.curated, this.props.search, this.props.featured).then(image => this.set(image));
+      this.fetchImage().then(image => this.setImage(image));
     }
 
     // Fetch next image and replace in cache
-    this.fetch(this.props.curated, this.props.search, this.props.featured).then(next => this.props.setLocal({ next }));
+    this.fetchImage().then(next => this.props.setLocal({ next }));
   }
 
   componentWillReceiveProps(nextProps: Props) {
     if (nextProps.featured !== this.props.featured || nextProps.curated !== this.props.curated) {
-      this.refresh(nextProps.curated, nextProps.search, nextProps.featured);
+      this.refreshDebounced.cancel();
+      this.refresh(nextProps);
     }
 
     if (nextProps.search !== this.props.search ) {
-      this.debouncedRefresh(nextProps.curated, nextProps.search, nextProps.featured);
+      this.refreshDebounced(nextProps);
     }
   }
 
@@ -79,26 +78,19 @@ class Unsplash extends React.PureComponent<Props, State> {
     );
   }
 
-  private refresh(curated?: boolean, search?: string, featured?: boolean) {
-    // Fetch current image
-    this.fetch(curated, search, featured).then(image => this.set(image));
-
-    // Clear and fetch next
-    this.props.setLocal({ next: undefined });
-    this.fetch(curated, search, featured).then(next => this.props.setLocal({ next }));
+  private refresh(props: Props) {
+    this.fetchImage(props).then(image => this.setImage(image));
+    this.fetchImage(props).then(next => this.props.setLocal({ next }));
   }
 
-  private async fetch(curated?: boolean, search?: string, featured?: boolean) {
-    const request = new Request(
-      'https://api.unsplash.com/photos/random?orientation=landscape' +
-      (curated
-        ? `&collections=${officialCollection}`
-        : ((featured ? '&featured=true' : '') + (search ? `&query=${search}` : ''))
-      ),
-      { headers: { Authorization: `Client-ID ${UNSPLASH_API_KEY}` } },
+  private async fetchImage({ curated, featured, search }: Props = this.props): Promise<Image> {
+    const headers = { Authorization: `Client-ID ${UNSPLASH_API_KEY}` };
+    const url = 'https://api.unsplash.com/photos/random?' + (curated
+      ? `collections=${officialCollection}`
+      : 'orientation=landscape' + (featured ? '&featured=true' : '') + (search ? `&query=${search}` : '')
     );
 
-    const res = await (await fetch(request)).json();
+    const res = await (await fetch(url, { headers })).json();
     const data = await (await fetch(res.urls.raw + '?q=85&w=1920')).blob();
 
     return {
@@ -110,13 +102,13 @@ class Unsplash extends React.PureComponent<Props, State> {
     };
   }
 
-  private set(image: Image) {
-    this.setState({
-      current: {
-        ...image,
-        src: URL.createObjectURL(image.data),
-      },
-    });
+  private setImage(image: Image) {
+    const current = {
+      ...image,
+      src: URL.createObjectURL(image.data),
+    };
+
+    this.setState({ current });
   }
 }